+++ Configuration/TCA/Overrides/tt_content.php
@@ -2,6 +2,7 @@
 
 use TYPO3\CMS\Core\Core\Environment;
 use TYPO3\CMS\Core\Information\Typo3Version;
+use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 // all use statements must come first
 //use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
@@ -16,10 +17,17 @@ defined('TYPO3') or die();
     /******************************************************************
      * FRONTEND PLUGINS - START
      *****************************************************************/
-    \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(
+    $pluginSignature = \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(
         'Feuserregistration',
         'Subscribe',
         'Feuserregistration: Display a email registration form'
+    );
+
+    ExtensionManagementUtility::addToAllTCAtypes(
+        'tt_content',
+        '--div--;Configuration,pages',
+        $pluginSignature,
+        'after:palette:headers'
     );
 
     \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(
@@ -29,7 +37,7 @@ defined('TYPO3') or die();
     );
 
     if (GeneralUtility::makeInstance(Typo3Version::class)->getMajorVersion() >= 14) {
-        \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(
+        $pluginSignature = \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(
             'Feuserregistration',
             'Verify',
             'Feuserregistration: Verify a registration.',
@@ -37,6 +45,13 @@ defined('TYPO3') or die();
             'plugins',
             '',
             'FILE:EXT:' . $extensionName . '/Configuration/Flexforms/Verify.xml'
+        );
+
+        ExtensionManagementUtility::addToAllTCAtypes(
+            'tt_content',
+            'pages',
+            $pluginSignature,
+            'after:pi_flexform'
         );
     } else {
         $pluginSignature = \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(